Allow editing label on click in main code editor screen

+++ navigator/views/codebook_views.py
@@ -265,7 +265,7 @@ class CodebookSaveChangesetsView(CodebookFormActionView):
             ccode.save(validate=False)
 
         # Check for any cycles. 
-        CodebookHierarchyResource.get_tree(Codebook.objects.get(id=codebook.id), include_labels=False)
+        CodebookHierarchyResource.get_tree(Codebook.objects.get(id=codebook.id))
 
 class CodebookAddCodeView(CodebookFormActionView):
     """
@@ -299,10 +299,10 @@ class CodebookSaveLabelsView(CodebookFormActionView):
     """
     url_fragment = "save-labels"
     class form_class(forms.Form):
+        code = JSONFormField()
         label = forms.CharField()
-        labels = JSONFormField()
+        labels = JSONFormField(required=False)
         parent = JSONFormField(required=False)
-        code = JSONFormField(required=False)
 
     def action(self, codebook, form):
         code = Code.objects.get(id=form.cleaned_data["code"])
@@ -312,30 +312,31 @@ class CodebookSaveLabelsView(CodebookFormActionView):
         if label:
             code.label = label
             code.save()
-        
-        # Get changed, deleted and new labels
-        changed_labels_map = { int(lbl.get("id")) : lbl for lbl in labels if lbl.get("id") is not None}
-        changed_labels = set(Label.objects.filter(id__in=changed_labels_map.keys()))
+            
+        if labels:
+            # Get changed, deleted and new labels
+            changed_labels_map = { int(lbl.get("id")) : lbl for lbl in labels if lbl.get("id") is not None}
+            changed_labels = set(Label.objects.filter(id__in=changed_labels_map.keys()))
 
-        created_labels = [Label(
-            language=Language.objects.get(id=lbl.get("language")), code=code, label=lbl["label"]
-        ) for lbl in labels if lbl.get("id", None) is None]
+            created_labels = [Label(
+                language=Language.objects.get(id=lbl.get("language")), code=code, label=lbl["label"]
+            ) for lbl in labels if lbl.get("id", None) is None]
 
-        current_labels = set(code.labels.all())
-        deleted_labels = current_labels - changed_labels
+            current_labels = set(code.labels.all())
+            deleted_labels = current_labels - changed_labels
 
-        # Deleting requested labels
-        for lbl in deleted_labels:
-            lbl.delete()
+            # Deleting requested labels
+            for lbl in deleted_labels:
+                lbl.delete()
 
-        # Create new labels
-        Label.objects.bulk_create(created_labels)
+            # Create new labels
+            Label.objects.bulk_create(created_labels)
 
-        # Update existing labels
-        for label in changed_labels:
-            label.language = Language.objects.get(id=changed_labels_map[label.id]["language"])
-            label.label = changed_labels_map[label.id]["label"]
-            label.save()
+            # Update existing labels
+            for label in changed_labels:
+                label.language = Language.objects.get(id=changed_labels_map[label.id]["language"])
+                label.label = changed_labels_map[label.id]["label"]
+                label.save()
 
         content = json.dumps(dict(code_id=code.id))
         return HttpResponse(content=content, status=201, content_type="application/json")
